fix: Remove walkdir in favor of an optimized implementation
WalkDir is a great library but we weren't using it very effectively and
when using it as intended it was much slower for our use case.

By switching to a simple read_dir based implementation we've made it
more maintainable code as well as shaved 50% off of the runtime (though
it was already so fast no one will really notice).